+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Seperated character randomization buttons into appearance and name.
+
 ### Removed
 
 ### Fixed